+++ lib/ash_introspection/rpc/errors.ex
@@ -97,7 +97,7 @@ defmodule AshIntrospection.Rpc.Errors do
         %{
           message: Exception.message(error),
           short_message: error.__struct__ |> Module.split() |> List.last(),
-          code: Macro.underscore(error.__struct__ |> Module.split() |> List.last()),
+          type: Macro.underscore(error.__struct__ |> Module.split() |> List.last()),
           vars: %{},
           fields: [],
           path: Map.get(error, :path, [])
@@ -187,7 +187,7 @@ defmodule AshIntrospection.Rpc.Errors do
     %{
       message: "Something went wrong. Unique error id: #{uuid}",
       short_message: "Internal error",
-      code: "internal_error",
+      type: "internal_error",
       vars: %{},
       fields: [],
       path: path,
@@ -244,7 +244,7 @@ defmodule AshIntrospection.Rpc.Errors do
         %{
           message: error.message,
           short_message: "Error description",
-          code: "error_code",
+          type: "error_type",
           vars: %{},
           fields: [],
           path: error.path || []
@@ -272,7 +272,7 @@ defmodule AshIntrospection.Rpc.Errors do
     %{
       message: "something went wrong",
       short_message: "Error",
-      code: "error",
+      type: "error",
       vars: %{},
       fields: [],
       path: Map.get(error, :path, [])
@@ -283,7 +283,7 @@ defmodule AshIntrospection.Rpc.Errors do
     %{
       message: "something went wrong",
       short_message: "Error",
-      code: "error",
+      type: "error",
       vars: %{},
       fields: [],
       path: []